+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -23,7 +30,7 @@ public class TaskList {
 
     public Task markTask(int index) throws DukeException {
         if (index < 0 || index >= this.size()) {
-            throw new DukeException("Task number does not exist.");
+            throw new DukeException("duke.task.Task number does not exist.");
         }
         Task task = this.tasks.get(index);
         task.mark();
@@ -32,7 +39,7 @@ public class TaskList {
 
     public Task unmarkTask(int index) throws DukeException {
         if (index < 0 || index >= this.size()) {
-            throw new DukeException("Task number does not exist.");
+            throw new DukeException("duke.task.Task number does not exist.");
         }
         Task task = this.tasks.get(index);
         task.unmark();
@@ -41,7 +48,7 @@ public class TaskList {
 
     public Task deleteTask(int index) throws DukeException {
         if (index < 0 || index >= this.size()) {
-            throw new DukeException("Task number does not exist.");
+            throw new DukeException("duke.task.Task number does not exist.");
         }
         Task task = this.tasks.remove(index);
         return task;
