+++ src/main/java/duke/command/ByeCommand.java
@@ -1,4 +1,11 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ByeCommand extends Command {
+
     @Override
     public boolean isExit() {
         return true;
