+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 import java.time.LocalDate;
 
 public class DeadlineCommand extends Command {
