+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,15 +1,21 @@
-public class DeleteCommand extends Command {
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class UnmarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(Storage storage, TaskList taskList, Ui ui) {
         try {
-            ui.printDeleteTask(taskList.deleteTask(index - 1));
-            ui.printSizeOfList(taskList.size());
+            ui.printUnmarkTask(taskList.unmarkTask(index - 1));
             storage.save(taskList.getTasks());
         } catch (DukeException e) {
             ui.printErrorMessage(e.getMessage());
